requirements.txt
django
djangorestframework
openai
python-dotenv